+++ users/serializers.py
@@ -1,5 +1,5 @@
 from rest_framework import serializers
-from django.contrib.auth import get_user_model
+from django.contrib.auth import get_user_model,authenticate
 from django.contrib.auth.password_validation import validate_password
 from django.db import transaction
 from wallets.models import Wallet
@@ -44,3 +44,23 @@ class RegisterSerializer(serializers.ModelSerializer):
         Wallet.objects.create(user=user)
 
         return user
+
+class LoginSerializer(serializers.Serializer):
+    email = serializers.EmailField()
+    password = serializers.CharField(write_only=True)
+
+    def validate(self, attrs):
+        user = authenticate(
+            email=attrs["email"],
+            password=attrs["password"]
+
+        )
+
+        if not user:
+            raise serializers.ValidationError("Invalid email or password")
+        
+        if not user.is_active:
+            raise serializers.ValidationError("User is disabled")
+        
+        attrs["user"] = user
+        return attrs